+++ cFS/apps/gsInterface/fsw/src/gsProcessMessages.c
@@ -521,7 +521,7 @@ void gsInterface_ProcessPacket() {
 
                 double heading = fmod(2*M_PI + atan2(pos->ve,pos->vn),2*M_PI)*180/M_PI;
                 double speed = sqrt(pos->vn*pos->vn + pos->ve*pos->ve);
-                mavlink_msg_adsb_vehicle_pack(sysid_ic,compid_ic,&msg,(uint8_t)pos->aircraft_id,
+                mavlink_msg_adsb_vehicle_pack(sysid_ic,compid_ic,&msg,pos->aircraft_id,
                                           (int32_t)(pos->latitude*1E7),
                                           (int32_t)(pos->longitude*1E7),
                                           ADSB_ALTITUDE_TYPE_GEOMETRIC,
@@ -573,7 +573,7 @@ void gsInterface_ProcessPacket() {
 
             double heading = fmod(2 * M_PI + atan2(traffic->ve, traffic->vn), 2 * M_PI) * 180 / M_PI;
             double speed = sqrt(traffic->vn * traffic->vn + traffic->ve * traffic->ve);
-            mavlink_msg_adsb_vehicle_pack(sysid_ic, compid_ic, &msg, (uint8_t)traffic->index,
+            mavlink_msg_adsb_vehicle_pack(sysid_ic, compid_ic, &msg, traffic->index,
                                           (int32_t)(traffic->latitude * 1E7),
                                           (int32_t)(traffic->longitude * 1E7),
                                           ADSB_ALTITUDE_TYPE_GEOMETRIC,
